quick fix

+++ python/TBAI/ai.py
@@ -83,7 +83,7 @@ class AIPlayer(Player):
             return (float(victor), 0)
         features = self._feature_extractor(state)
         features = torch.from_numpy(state.features())
-        features = features.type(torch.DoubleTensor)
+        features = features.type(torch.FloatTensor)
         features = Variable(features)
         # PENDING: use neural net
         if self._model:
